server_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-telegram-bot==20.7
slowapi==0.1.9
cryptography==41.0.7

